all-in-one/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
databases[postgresql]==0.8.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# Utils
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4

# Email
jinja2==3.1.2
emails==0.6

# File Upload
pillow==10.1.0
python-magic==0.4.27

# Maps & Geo
geopy==2.4.0
haversine==2.8.0
openrouteservice==2.3.1

# ML/Recommendations
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.3

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Async
httpx==0.25.1
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1